feat: SpringApplication에서 controller 못읽는 문제 해결
