+++ build.gradle
@@ -1,119 +1,43 @@
-//plugins {
-//    id 'java'
-//    id 'idea'
-//    id 'org.springframework.boot' version '2.7.9'
-//    id 'io.spring.dependency-management' version '1.0.15.RELEASE'
-//}
-//
-//group 'codesquad'
-//version '0.0.1-SNAPSHOT'
-//sourceCompatibility = 11
-//
-//configurations {
-//    compileOnly {
-//        extendsFrom annotationProcessor
-//    }
-//}
-//
-//repositories {
-//    mavenCentral()
-//    maven { url 'https://repo.spring.io/milestone' }
-//    maven { url 'https://repo.spring.io/snapshot' }
-//}
-//
-//dependencies {
-////    implementation 'org.springframework.boot:spring-boot-starter-web'
-////    maven 'org.springframework.boot:spring-boot-starter-web:2.4.0'
-//    //spring
-//    implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
-//    implementation 'org.springframework.boot:spring-boot-starter-web'
-//    compileOnly 'org.projectlombok:lombok'
-//    annotationProcessor 'org.projectlombok:lombok'
-//    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-//
-//    // 기존 설정
-//    implementation 'ch.qos.logback:logback-classic:1.2.3'
-//    implementation 'com.google.guava:guava:29.0-jre'
-//    implementation 'ch.qos.logback:logback-classic:1.2.3'
-//    implementation 'com.github.jknack:handlebars:4.2.0'
-//
-//    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
-//    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
-//    testImplementation 'org.assertj:assertj-core:3.16.1'
-//}
-//
-//tasks.named('test') {
-//    useJUnitPlatform()
-//}
-//
-
-buildscript {
-    repositories {
-        mavenLocal()
-        maven { url 'https://maven.aliyun.com/repository/google/' }
-        maven { url 'https://maven.aliyun.com/repository/public/' }
-        maven { url 'https://maven.aliyun.com/repository/spring/' }
-        maven { url 'https://maven.aliyun.com/repository/gradle-plugin/' }
-        maven { url 'https://maven.aliyun.com/repository/spring-plugin/' }
-        maven {
-            url "https://plugins.gradle.org/m2/"
-        }
-        mavenCentral()
-    }
-    dependencies {
-        classpath("org.springframework.boot:spring-boot-gradle-plugin:2.5.6")
-        classpath "io.spring.gradle:dependency-management-plugin:1.0.11.RELEASE"
-    }
-}
-
 plugins {
-    id 'java'
+	id 'java'
+	id 'org.springframework.boot' version '2.7.9-SNAPSHOT'
+	id 'io.spring.dependency-management' version '1.0.15.RELEASE'
 }
 
-apply plugin: "org.springframework.boot"
-apply plugin: "io.spring.dependency-management"
-
-//plugins {
-//    id 'java'
-//    id 'org.springframework.boot' version '3.0.2'
-//    id 'io.spring.dependency-management' version '1.1.0'
-//}
-
-group = 'Board'
+group = 'com.example'
 version = '0.0.1-SNAPSHOT'
 sourceCompatibility = '11'
 
 configurations {
-    compileOnly {
-        extendsFrom annotationProcessor
-    }
+	compileOnly {
+		extendsFrom annotationProcessor
+	}
 }
 
 repositories {
-    mavenCentral()
+	mavenCentral()
+	maven { url 'https://repo.spring.io/milestone' }
+	maven { url 'https://repo.spring.io/snapshot' }
 }
 
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    compileOnly 'org.projectlombok:lombok'
-//    runtimeOnly 'com.mysql:mysql-connector-j'
-    annotationProcessor 'org.projectlombok:lombok'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+	implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
+	implementation 'org.springframework.boot:spring-boot-starter-web'
+	compileOnly 'org.projectlombok:lombok'
+	annotationProcessor 'org.projectlombok:lombok'
+	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 
-    // 기존 설정
-    implementation 'ch.qos.logback:logback-classic:1.2.3'
-    implementation 'com.google.guava:guava:29.0-jre'
-    implementation 'ch.qos.logback:logback-classic:1.2.3'
-    implementation 'com.github.jknack:handlebars:4.2.0'
+	// 기존 설정
+	implementation 'ch.qos.logback:logback-classic:1.2.3'
+	implementation 'com.google.guava:guava:29.0-jre'
+	implementation 'ch.qos.logback:logback-classic:1.2.3'
+	implementation 'com.github.jknack:handlebars:4.2.0'
 
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
-    testImplementation 'org.assertj:assertj-core:3.16.1'
+	testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
+	testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
+	testImplementation 'org.assertj:assertj-core:3.16.1'
 }
 
 tasks.named('test') {
-    useJUnitPlatform()
+	useJUnitPlatform()
 }
-
-
