+++ settings.gradle
@@ -1,2 +1,8 @@
-rootProject.name = 'java-was-2022'
-
+pluginManagement {
+	repositories {
+		maven { url 'https://repo.spring.io/milestone' }
+		maven { url 'https://repo.spring.io/snapshot' }
+		gradlePluginPortal()
+	}
+}
+rootProject.name = 'demo'
