+++ src/main/java/com/example/demo/config/MvcConfiguration.java
@@ -1,0 +1,19 @@
+package com.example.demo.config;
+
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+//@Configuration
+public class MvcConfiguration implements WebMvcConfigurer {
+
+//    private static final String[] CLASSPATH_RESOURCE_LOCATIONS = {
+//            "classpath:/META-INF/resources/", "classpath:/resources/", "classpath:/templates/",
+//            "classpath:/static/", "classpath:/public/"
+//    };
+//
+//    @Override
+//    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+//        registry.addResourceHandler("/**")
+//                .addResourceLocations(CLASSPATH_RESOURCE_LOCATIONS)
+//                .setCacheControl(CacheControl.maxAge(10, TimeUnit.MINUTES));
+//    }
+}
