+++ src/main/java/com/example/demo/controller/MainController.java
@@ -1,0 +1,16 @@
+package com.example.demo.controller;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Slf4j
+@Controller
+public class MainController {
+
+    @RequestMapping("/index")
+    public String mainPage(){
+        return "index";
+    }
+
+}
