+++ src/main/java/com/example/demo/service/userService.java
@@ -1,0 +1,10 @@
+package com.example.demo.service;
+
+public class userService {
+
+    public void join(String userName, String userId,String password,String email) {
+
+//        User user = new User();
+
+    }
+}
