--- src/main/java/config/MvcConfiguration.java
@@ -1,30 +1,0 @@
-package config;
-
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.CacheControl;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.concurrent.TimeUnit;
-
-@Configuration
-@EnableAutoConfiguration(exclude={DataSourceAutoConfiguration.class})
-public class MvcConfiguration implements WebMvcConfigurer {
-
-//    <!--<script th:src="@{../static/js/jquery-2.2.0.min.js}"></script>-->
-
-    private static final String[] CLASSPATH_RESOURCE_LOCATIONS = {
-            "classpath:/META-INF/resources/", "classpath:/resources/",
-            "classpath:/static/", "classpath:/public/",
-            "classpath:/html5shim.googlecode.com/"
-    };
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/**")
-                .addResourceLocations(CLASSPATH_RESOURCE_LOCATIONS)
-                .setCacheControl(CacheControl.maxAge(10, TimeUnit.MINUTES));
-    }
-}
