--- src/main/java/controller/MainController.java
@@ -1,14 +1,0 @@
-package controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class MainController {
-
-    @GetMapping("/")
-    public String main(){
-        return "index";
-    }
-
-}
