--- src/main/java/controller/UserController.java
@@ -1,7 +1,0 @@
-package controller;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class UserController {
-}
