--- src/main/java/webserver/BoardApplication.java
@@ -1,11 +1,0 @@
-package webserver;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BoardApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(BoardApplication.class, args);
-    }
-}
