+++ src/main/resources/templates/index.html
@@ -1,160 +1,10 @@
 <!DOCTYPE html>
-<html lang="kr" xmlns:th="http://www.thymeleaf.org">
-	<head>
-		<meta http-equiv="content-type" content="text/html; charset=UTF-8">
-		<meta charset="utf-8">
-		<title>SLiPP Java Web Programming</title>
-		<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Title</title>
+</head>
+<body>
 
-        <!--[if lt IE 9]>
-            <script src="//html5shim.googlecode.com/svn/trunk/html5.js"></script>
-        <![endif]-->
-
-        <link th:href="@{../css/bootstrap.min.css}" rel="stylesheet">
-		<link th:href="@{../css/styles.css}" rel="stylesheet">
-	</head>
-	
-	<body>
-<nav class="navbar navbar-fixed-top header">
- 	<div class="col-md-12">
-        <div class="navbar-header">
-
-            <a href="index.html" class="navbar-brand">SLiPP</a>
-          <button type="button" class="navbar-toggle" data-toggle="collapse" data-target="#navbar-collapse1">
-          <i class="glyphicon glyphicon-search"></i>
-          </button>
-      
-        </div>
-        <div class="collapse navbar-collapse" id="navbar-collapse1">
-          <form class="navbar-form pull-left">
-              <div class="input-group" style="max-width:470px;">
-                <input type="text" class="form-control" placeholder="Search" name="srch-term" id="srch-term">
-                <div class="input-group-btn">
-                  <button class="btn btn-default btn-primary" type="submit"><i class="glyphicon glyphicon-search"></i></button>
-                </div>
-              </div>
-          </form>
-          <ul class="nav navbar-nav navbar-right">             
-             <li>
-                <a href="#" class="dropdown-toggle" data-toggle="dropdown"><i class="glyphicon glyphicon-bell"></i></a>
-                <ul class="dropdown-menu">
-                  <li><a href="https://slipp.net" target="_blank">SLiPP</a></li>
-                  <li><a href="https://facebook.com" target="_blank">Facebook</a></li>
-                </ul>
-             </li>
-             <li><a href="user/list.html"><i class="glyphicon glyphicon-user"></i></a></li>
-           </ul>
-        </div>	
-     </div>	
-</nav>
-<div class="navbar navbar-default" id="subnav">
-    <div class="col-md-12">
-        <div class="navbar-header">
-            <a href="#" style="margin-left:15px;" class="navbar-btn btn btn-default btn-plus dropdown-toggle" data-toggle="dropdown"><i class="glyphicon glyphicon-home" style="color:#dd1111;"></i> Home <small><i class="glyphicon glyphicon-chevron-down"></i></small></a>
-            <ul class="nav dropdown-menu">
-<!--                <li><a href="user/profile.html"><i class="glyphicon glyphicon-user" style="color:#1111dd;"></i> Profile</a></li>-->
-                <li><a th:href="@{user/profile.html}"><i class="glyphicon glyphicon-user" style="color:#1111dd;"></i> Profile</a></li>
-                <li class="nav-divider"></li>
-                <li><a href="#"><i class="glyphicon glyphicon-cog" style="color:#dd1111;"></i> Settings</a></li>
-            </ul>
-            
-            <button type="button" class="navbar-toggle" data-toggle="collapse" data-target="#navbar-collapse2">
-            	<span class="sr-only">Toggle navigation</span>
-            	<span class="icon-bar"></span>
-            	<span class="icon-bar"></span>
-            	<span class="icon-bar"></span>
-            </button>            
-        </div>
-        <div class="collapse navbar-collapse" id="navbar-collapse2">
-            <ul class="nav navbar-nav navbar-right">
-                <li class="active"><a href="index.html">Posts</a></li>
-<!--                <li><a href="user/login.html" role="button">로그인</a></li>-->
-                <li><a th:href="@{user/login.html}" role="button">로그인</a></li>
-<!--                <li><a href="user/form.html" role="button">회원가입</a></li>-->
-                <li><a th:href="@{user/form.html}" role="button">회원가입</a></li>
-                <!--
-                <li><a href="#loginModal" role="button" data-toggle="modal">로그인</a></li>
-                <li><a href="#registerModal" role="button" data-toggle="modal">회원가입</a></li>
-                -->
-                <li><a href="#" role="button">로그아웃</a></li>
-                <li><a href="#" role="button">개인정보수정</a></li>
-            </ul>
-        </div>
-    </div>
-</div>
-
-<div class="container" id="main">
-   <div class="col-md-12 col-sm-12 col-lg-10 col-lg-offset-1">
-      <div class="panel panel-default qna-list">
-          <ul class="list">
-              <li>
-                  <div class="wrap">
-                      <div class="main">
-                          <strong class="subject">
-                              <a href="qna/show.html">국내에서 Ruby on Rails와 Play가 활성화되기 힘든 이유는 뭘까?</a>
-                          </strong>
-                          <div class="auth-info">
-                              <i class="icon-add-comment"></i>
-                              <span class="time">2016-01-15 18:47</span>
-<!--                              <a href="user/profile.html" class="author">자바지기</a>-->
-                              <a th:href="@{user/profile.html}" class="author">자바지기</a>
-                          </div>
-                          <div class="reply" title="댓글">
-                              <i class="icon-reply"></i>
-                              <span class="point">8</span>
-                          </div>
-                      </div>
-                  </div>
-              </li>
-              <li>
-                  <div class="wrap">
-                      <div class="main">
-                          <strong class="subject">
-                              <a href="qna/show.html">runtime 에 reflect 발동 주체 객체가 뭔지 알 방법이 있을까요?</a>
-                          </strong>
-                          <div class="auth-info">
-                              <i class="icon-add-comment"></i>
-                              <span class="time">2016-01-05 18:47</span>
-                              <a href="user/profile.html" class="author">김문수</a>
-                              <a th:href="@{user/profile.html}" class="author">김문수</a>
-                          </div>
-                          <div class="reply" title="댓글">
-                              <i class="icon-reply"></i>
-                              <span class="point">12</span>
-                          </div>
-                      </div>
-                  </div>
-              </li>
-          </ul>
-          <div class="row">
-              <div class="col-md-3"></div>
-              <div class="col-md-6 text-center">
-                  <ul class="pagination center-block" style="display:inline-block;">
-                      <li><a href="#">«</a></li>
-                      <li><a href="#">1</a></li>
-                      <li><a href="#">2</a></li>
-                      <li><a href="#">3</a></li>
-                      <li><a href="#">4</a></li>
-                      <li><a href="#">5</a></li>
-                      <li><a href="#">»</a></li>
-                </ul>
-              </div>
-              <div class="col-md-3 qna-write">
-<!--                  <a href="qna/form.html" class="btn btn-primary pull-right" role="button">질문하기</a>-->
-                  <a th:href="@{qna/form.html}" class="btn btn-primary pull-right" role="button">질문하기</a>
-              </div>
-          </div>
-        </div>
-    </div>
-</div>
-
-<!-- script references -->
-<script th:src="@{../js/jquery-2.2.0.min.js}"></script>
-<script th:src="@{../js/bootstrap.min.js}"></script>
-<script th:src="@{../js/scripts.js}"></script>
-
-<!--<script src="../js/jquery-2.2.0.min.js"></script>-->
-<!--<script src="../js/bootstrap.min.js"></script>-->
-<!--<script src="../js/scripts.js"></script>-->
-	</body>
+</body>
 </html>
