+++ src/main/resources/templates/user/login.html
@@ -1,15 +1,17 @@
 <!DOCTYPE html>
-<html lang="kr">
+<html lang="kr" xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta http-equiv="content-type" content="text/html; charset=UTF-8">
     <meta charset="utf-8">
     <title>SLiPP Java Web Programming</title>
     <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
-    <link href="../css/bootstrap.min.css" rel="stylesheet">
+<!--    <link href="../css/bootstrap.min.css" rel="stylesheet">-->
+    <link th:href="@{../css/bootstrap.min.css}" rel="stylesheet">
     <!--[if lt IE 9]>
     <script src="//html5shim.googlecode.com/svn/trunk/html5.js"></script>
     <![endif]-->
-    <link href="../css/styles.css" rel="stylesheet">
+    <link th:href="@{../css/styles.css}" rel="stylesheet">
+<!--    <link href="../css/styles.css" rel="stylesheet">-->
 </head>
 <body>
 <nav class="navbar navbar-fixed-top header">
