--- src/test/resources/postResponse.txt
@@ -1,8 +1,0 @@
-POST /user/create HTTP/1.1
-Host: localhost:8080
-Connection: keep-alive
-Content-Length: 93
-Content-Type: application/x-www-form-urlencoded
-Accept: */*
-
-userId=javajigi&password=password&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net
